*hannes: format
git-svn-id: file:///usr/local/Singular/svn/trunk@11068 2c84dea3-7e68-4137-9b89-c4e89433aadc

+++ Singular/feOpt.cc
@@ -1,7 +1,7 @@
 /****************************************
 *  Computer Algebra System SINGULAR     *
 ****************************************/
-/* $Id: feOpt.cc,v 1.23 2008-09-10 09:15:52 Singular Exp $ */
+/* $Id: feOpt.cc,v 1.24 2008-09-22 17:24:03 Singular Exp $ */
 /*
 * ABSTRACT: Implementation of option buisness
 */
@@ -491,7 +491,7 @@ void fePrintOptValues()
       }
       else
       {
-        Print("// --%-15s %d\n", feOptSpec[i].name, feOptSpec[i].value);
+        Print("// --%-15s %d\n", feOptSpec[i].name, (int)(long)feOptSpec[i].value);
       }
     }
     i++;
